app/requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.2
numpy==1.22.3
python-dotenv==0.20.0
requests==2.22.0
transliterate==1.10.2
Werkzeug==2.1.2
waitress